Move ProxImaL to contrib
